Move cmscommon et al under the cms namespace
This commit essentially is just:

- git mv cmscommon cms/common
- git mv cmscontrib cms/contrib
- git mv cmstaskenv cms/taskenv
- git mv cmsranking cms/ranking
- git mv cmstestsuite cms/testsuite

This change is not strictly necessary while using setup.py because `find_packages()` anyway looks for all folders containing a `__init__.py` file so it installs all of these "unrelated" packages anyway.

With tools such as poetry, however, having a single package for CMS makes things easier, as poetry installs the root package (detected as "cms") in editable mode and it doesn't install the other packages.